+++ source/nand/nand.h
@@ -18,6 +18,6 @@ int ReadNandSectors(u8* buffer, u32 sector, u32 count, u32 keyslot, u32 src);
 int WriteNandSectors(const u8* buffer, u32 sector, u32 count, u32 keyslot, u32 dest);
 
 u64 GetNandSizeSectors(u32 src);
-u8 CheckNandType(u32 src);
+u32 CheckNandType(u32 src);
 
 bool InitEmuNandBase(void);
